requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
Naked==0.1.32
pyasn1==0.5.1
pycryptodome==3.20.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shellescape==3.8.1
urllib3==2.1.0